Fix for: "secstore total" returns error
The problem is, that "GetTotal()" API is not implemented in
the TrustZone-based secure storage. The reason being, the data
is saved on RW flash filesystem, and the cap is not enforced.

The fix is to introduce total secure storage limit of 1Mb;
although it's not enforced, this will allow to return some
meaningful numbers for total space and free space.

Resolves: LE-13169
Change-Id: I2a090286c54da8ada9524241353b192d16545d5f
(cherry picked from commit de47b2bcafeaea05137430537f17d0df3564b395)