Fix for secStore Write() failing after key is re-created.
File with TrustZone keys can be lost for example due to Legato
being temporarily downgraded and then upgraded again to 19.07+.
In this case, the total size calculation is failing, and this
is preventing the Write() operation from succeeding. As a
result, the old data can't be overwritten.

The fix is to make size calculation tolerant to decryption failure.
If the data can't be decrypted, just use the encrypted size
(which is slightly larger) and proceed.

Resolves: LE-13342
Change-Id: Ia64635ae15badbabebd43d4a9531bfa021dc4e93