Fix for "secstore readmeta" command not working.
The problem is that in new TrustZone-based implementation, the meta
file does not exist, and can't be simply copied to the provided path.

The fix is to generate the meta data on demand, and write it to the
provided file path.

Resolves: LE-13343
Change-Id: I9acc0f890da6bcd09369ed5098730d324c152e39
(cherry picked from commit e8c502e3cda0ba3515e18faa9450431bc3d752bc)