Fixes connecting to AVC when using IOTKeyStore
Issue:
Device cannot connect to the AVC server when IOT keystore is
enabled because it cannot find credentials. The PA uses SFS paths
to look for credentials that are saved to the EFS, so it's unable
to find them.

Solution:
Convert SFS paths to a generic path that is different between
RC76 and WP products.

Resolves: LE-15788
Change-Id: Ib06c2952a99d722b2e8c047c142f42d6fa232d14