Make secure storage backwards compatible
Currently, secure storage determines which TZ to use based off the build time flags. However, if the device does not support the specified TZ, the secure storage will assert and fail. This change will also check at runtime whether the specific build flags are applicable. If not, it will fallback to the working version.

Resolves: LE-16386
Change-Id: I523f88bbf7c5a7e112985f9e8a048c1e87d19912